cleanup types
