+++ frontend-admin/src/api/types.ts
@@ -1,0 +1,14 @@
+export interface Van {
+  id: number;
+  routeId: number;
+}
+export interface Route {
+  id: number;
+  name: string;
+  description: string;
+}
+
+export interface Stop {
+  id: number;
+  name: string;
+}
