+++ frontend-admin/src/pages/routes/route-types.tsx
@@ -1,13 +1,4 @@
-export interface Route {
-  id: number;
-  name: string;
-  description: string;
-}
-
-export interface Stop {
-  id: number;
-  name: string;
-}
+import { Route } from "../../api/types";
 
 export interface AddRouteFormProps {
   onSubmit: (data: FormData) => void; // or Promise<void> if async
