--- frontend-admin/src/pages/vans/van-types.tsx
@@ -1,4 +1,0 @@
-export interface Van {
-  id: number;
-  routeId: number;
-}
